Videos controllers finished, lets start with comments

+++ server/src/controllers/videos.controllers.ts
@@ -2,6 +2,8 @@ import { RequestHandler } from "express";
 import Video from '../models/Video.js'
 import { error } from "../middlewares/error.js";
 import User from '../models/User.js'
+import { giveOneUser } from "./whatUser.js";
+
 
 export const getVideos: RequestHandler = async (req, res) => {
     const videos = await Video.find()
@@ -15,11 +17,11 @@ export const createVideo: RequestHandler = async (req, res) => {
         error({statusCode: 301, message: 'The video already exists please change the URL'}, res)
         return
     }
-    const findUser = await User.findOne({username: req.body.username})
-    if (!findUser) {
-        error({statusCode: 301, message: 'This user no exists please create one'}, res)
-        return
-    }
+    const recuperateUser = giveOneUser(req) 
+    const findUser = await User.findOne({username: recuperateUser.username})
+
+    if (!findUser) return error({statusCode: 401, message: "You have to register before create a video"}, res)
+
     try {
         const createNewVideo = new Video({title: req.body.title, description: req.body.description, url: req.body.url, userId: findUser})
         const savedVideo = await createNewVideo.save()
@@ -38,13 +40,27 @@ export const getVideo: RequestHandler = async (req, res) => {
 }
 
 export const updateVideo: RequestHandler = async (req, res) => {
+    const recuperateUserName = giveOneUser(req)
+    const recuperateUserId = await User.findOne({username: recuperateUserName.username})
+    const recuperateVideo = await Video.findById(req.params.id)
+    
+    if (!recuperateUserId || !recuperateVideo) return error({statusCode: 401, message: 'Error on user authorization'}, res)
+    if (!recuperateVideo.userId.equals(recuperateUserId._id)) return error({statusCode: 401, message: 'You cannot modify this video'}, res)
+    
     const videoUpdated = await Video.findByIdAndUpdate(req.params.id, req.body, {new: true})
     if (!videoUpdated) return error({statusCode: 204, message: `Any video with id: ${req.params.id} found`}, res)
     res.json({success: true, newVideo: `[*] ${videoUpdated}`})
 }
 
 export const deleteVideo: RequestHandler =  async (req, res) => {
+    const recuperateUserName = giveOneUser(req)
+    const recuperateUserId = await User.findOne({username: recuperateUserName.username})
+    const recuperateVideo = await Video.findById(req.params.id)
+    
+    if (!recuperateUserId || !recuperateVideo) return error({statusCode: 401, message: 'Error on user authorization'}, res)
+    if (!recuperateVideo.userId.equals(recuperateUserId._id)) return error({statusCode: 401, message: 'You cannot modify this video'}, res)
+    
     const videoFound = await Video.findByIdAndDelete(req.params.id)
     if (!videoFound) return error({statusCode: 204, message: `Any video with id: ${req.params.id} found`}, res)
     res.json({ success: true, video: `[*] ${videoFound}` })
-}
+}
